Fixing deprecated methods

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -1,18 +1,24 @@
 package br.com.alura.forum.controller;
 
+import br.com.alura.forum.dto.DetalhesTopicoDto;
 import br.com.alura.forum.dto.TopicoDto;
+import br.com.alura.forum.form.AtualizacaoTopicoForm;
 import br.com.alura.forum.modelo.Topico;
 import br.com.alura.forum.repository.CursoRepository;
 import br.com.alura.forum.repository.TopicoRepository;
 import br.com.alura.forum.form.TopicoForm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @RestController
 @RequestMapping("/topicos")
@@ -26,8 +32,8 @@ public class TopicosController {
 
 
     @GetMapping
-    public List<TopicoDto> lista(String nomeCurso){
-
+    public List<TopicoDto> lista(String nomeCurso){ // Esse parâmetro deve ser passado igual ao nome da variável:
+                                                    // http://localhost:8080/topicos?nomeCurso=Spring+Boot
         if(nomeCurso == null){
             List<Topico> topicos = topicoRepository.findAll();
             return TopicoDto.converter(topicos);
@@ -38,6 +44,7 @@ public class TopicosController {
     }
 
     @PostMapping
+    @Transactional // Avisar ao Spring pra comitar ao final do método
     public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder){
         Topico topico = form.converter(cursoRepository);
         topicoRepository.save(topico);
@@ -45,9 +52,42 @@ public class TopicosController {
         return ResponseEntity.created(uri).body(new TopicoDto(topico));
     }
 
+    //Parâmetro dinâmico
+    //Variável do Path, da url
+    @GetMapping("/{id}")
+    public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Long id){
+
+        Optional<Topico> topico = topicoRepository.findById(id); // findById retorna um Optional e não um Exception
+
+        if(topico.isPresent()){
+            return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));
+        }
+        return ResponseEntity.notFound().build();
+    }
+
+    @PutMapping("/{id}")
+    @Transactional // Avisar ao Spring pra comitar ao final do método
+    public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form){
+        Optional<Topico> optional = topicoRepository.findById(id);
+
+        if(optional.isPresent()){
+            Topico topico = form.atualizar(id, topicoRepository);
+            return ResponseEntity.ok(new TopicoDto(topico));
+        }
+
+        return ResponseEntity.notFound().build();
+    }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity deletar(@PathVariable("id") Long id){
-        topicoRepository.deleteById(id);
-        return ResponseEntity.noContent().build();
+    @Transactional // Avisar ao Spring pra comitar ao final do método
+    public ResponseEntity remover(@PathVariable Long id){
+        Optional<Topico> optional = topicoRepository.findById(id);
+
+        if(optional.isPresent()){
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+
+        return ResponseEntity.notFound().build();
     }
 }
